outsource freind list component
