+++ app/components/FriendList.tsx
@@ -1,25 +1,60 @@
-import React, { useEffect, useState } from 'react';
-import { View, Text, StyleSheet } from 'react-native';
-import { Button, Divider, SearchBar } from '@rneui/themed';
 import { generateClient, SelectionSet } from 'aws-amplify/api';
+import { useEffect, useState } from 'react';
+import { StyleSheet, Text, View } from 'react-native';
 import { Schema } from '../../amplify/data/resource';
+import { Button } from '@rneui/themed';
+
+const selectionSet = ['id', 'userIdOne', 'userIdTwo', 'isConfirmed'] as const;
+type Friends = SelectionSet<Schema['Friends']['type'], typeof selectionSet>;
 
 const client = generateClient<Schema>();
 
-const selectionSet = ['id', 'userIdOne', 'userIdTwo', 'isConfirmed'] as const;
-const selectionSetPlayer = ['id', 'name'] as const;
-type Friends = SelectionSet<Schema['Friends']['type'], typeof selectionSet>;
-type Player = SelectionSet<Schema['Player']['type'], typeof selectionSetPlayer>;
-
-export default function FriendListScreen({
+export default function FriendList({
   localPlayerId,
 }: {
   localPlayerId: string;
 }) {
   const [friends, setFriends] = useState<Friends[]>([]);
-  const [search, setSearch] = useState('');
-  const [searchResults, setSearchResults] = useState<Player[]>([]);
-  const [userNames, setUserNames] = useState<{ [key: string]: string }>({}); // State to hold userId to name mapping
+  const [userNames, setUserNames] = useState<{ [key: string]: string }>({});
+
+  const fetchUserNames = async (userIds: (string | null | undefined)[]) => {
+    // Filter out null or undefined values
+    const validUserIds = userIds.filter(
+      (id): id is string => id !== null && id !== undefined
+    );
+
+    const nameMapping: { [key: string]: string } = {};
+    for (const userId of validUserIds) {
+      const { data: player, errors } = await client.models.Player.get({
+        id: userId,
+      });
+      if (errors) {
+        console.error(errors);
+      } else if (player) {
+        // Ensure player.name is a string, use a fallback if it's null or undefined
+        nameMapping[userId] = player.name ?? 'Unknown Player';
+      }
+    }
+    setUserNames((prevState) => ({ ...prevState, ...nameMapping }));
+  };
+
+  const confirmFriend = async (friendId: string) => {
+    // only let second player (the one who was invited) confirm the friendship
+    if (
+      friends.find((friend) => friend.id === friendId)?.userIdTwo !==
+      localPlayerId
+    ) {
+      return;
+    }
+    await client.models.Friends.update({
+      id: friendId,
+      isConfirmed: true,
+    });
+  };
+
+  const removeFriend = async (friendId: string) => {
+    await client.models.Friends.delete({ id: friendId });
+  };
 
   useEffect(() => {
     if (!localPlayerId) return;
@@ -50,104 +85,8 @@ export default function FriendListScreen({
     return () => sub.unsubscribe();
   }, [localPlayerId]);
 
-  const fetchUserNames = async (userIds: (string | null | undefined)[]) => {
-    // Filter out null or undefined values
-    const validUserIds = userIds.filter(
-      (id): id is string => id !== null && id !== undefined
-    );
-
-    const nameMapping: { [key: string]: string } = {};
-    for (const userId of validUserIds) {
-      const { data: player, errors } = await client.models.Player.get({
-        id: userId,
-      });
-      if (errors) {
-        console.error(errors);
-      } else if (player) {
-        // Ensure player.name is a string, use a fallback if it's null or undefined
-        nameMapping[userId] = player.name ?? 'Unknown Player';
-      }
-    }
-    setUserNames((prevState) => ({ ...prevState, ...nameMapping }));
-  };
-
-  const searchPlayers = async (search: string) => {
-    const { data: player, errors } = await client.models.Player.list({
-      filter: {
-        name: {
-          contains: search,
-        },
-      },
-      selectionSet: selectionSetPlayer,
-    });
-    if (errors) {
-      console.error(errors);
-    } else {
-      setSearchResults(player);
-    }
-  };
-
-  const handleSearchTyping = (search: string) => {
-    setSearch(search);
-    if (search.length > 0) {
-      searchPlayers(search);
-    }
-  };
-
-  const handleInvite = async (playerId: string) => {
-    await client.models.Friends.create({
-      userIdOne: localPlayerId,
-      userIdTwo: playerId,
-      isConfirmed: false,
-    });
-  };
-
-  const confirmFriend = async (friendId: string) => {
-    // only let second player (the one who was invited) confirm the friendship
-    if (
-      friends.find((friend) => friend.id === friendId)?.userIdTwo !==
-      localPlayerId
-    ) {
-      return;
-    }
-    await client.models.Friends.update({
-      id: friendId,
-      isConfirmed: true,
-    });
-  };
-
-  const removeFriend = async (friendId: string) => {
-    await client.models.Friends.delete({ id: friendId });
-  };
-
   return (
-    <View style={styles.container}>
-      <SearchBar
-        placeholder='Type Here...'
-        onChangeText={handleSearchTyping}
-        value={search}
-        style={{ borderRadius: 10 }}
-      />
-      {searchResults.map((player) => {
-        return (
-          <View key={player.id} style={styles.gameItem}>
-            <Text style={styles.gameText}>{player.name}</Text>
-            <Button
-              title={'Invite'}
-              buttonStyle={styles.inviteButton}
-              onPress={() => handleInvite(player.id)}
-            />
-          </View>
-        );
-      })}
-      <Divider
-        style={{
-          backgroundColor: 'black',
-          height: 2,
-          marginVertical: 5,
-          opacity: 0.5,
-        }}
-      />
+    <View>
       {friends.map((friend) => {
         // Determine the correct friend ID
         const friendId =
